requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
pyyaml